Rework sub-modules
