+++ src/byte_writer.rs
