+++ src/internal/ppmd7.rs
@@ -1,3 +1,9 @@
+mod decoder;
+mod encoder;
+
+pub use decoder::*;
+pub use encoder::*;
+
 use super::*;
 
 #[derive(Copy, Clone)]
