+++ src/internal/ppmd8/decoder.rs
@@ -1,40 +1,48 @@
-use super::ppmd8::*;
 use super::*;
 
-pub unsafe fn range_encoder_flush(p: *mut Ppmd8) {
+pub unsafe fn range_decoder_init(p: *mut Ppmd8) -> i32 {
     let mut i: std::ffi::c_uint = 0;
+    (*p).code = 0 as std::ffi::c_int as u32;
+    (*p).range = 0xFFFFFFFF as std::ffi::c_uint;
+    (*p).low = 0 as std::ffi::c_int as u32;
     i = 0 as std::ffi::c_int as std::ffi::c_uint;
     while i < 4 as std::ffi::c_int as std::ffi::c_uint {
-        ((*(*p).stream.output).write).expect("non-null function pointer")(
-            (*p).stream.output,
-            ((*p).low >> 24 as std::ffi::c_int) as u8,
-        );
+        (*p).code = (*p).code << 8 as std::ffi::c_int
+            | ((*(*p).stream.input).read).expect("non-null function pointer")((*p).stream.input)
+                as u32;
         i = i.wrapping_add(1);
         i;
-        (*p).low <<= 8 as std::ffi::c_int;
     }
+    return ((*p).code < 0xFFFFFFFF as std::ffi::c_uint) as std::ffi::c_int;
 }
 
 #[inline(always)]
-unsafe fn range_encoder_encode(p: *mut Ppmd8, start: u32, size: u32, total: u32) {
-    (*p).range = (*p).range / total;
-    (*p).low = ((*p).low).wrapping_add(start * (*p).range);
+unsafe fn range_decoder_decode(p: *mut Ppmd8, mut start: u32, size: u32) {
+    start = start * (*p).range;
+    (*p).low = ((*p).low).wrapping_add(start);
+    (*p).code = ((*p).code).wrapping_sub(start);
     (*p).range = (*p).range * size;
 }
 
-pub unsafe fn encode_symbol(p: *mut Ppmd8, symbol: std::ffi::c_int) {
+pub unsafe fn decode_symbol(p: *mut Ppmd8) -> std::ffi::c_int {
     let mut charMask: [usize; 32] = [0; 32];
     if (*(*p).min_context).num_stats as std::ffi::c_int != 0 as std::ffi::c_int {
         let mut s: *mut State = ((*p).base).offset((*(*p).min_context).union4.stats as isize)
             as *mut std::ffi::c_void as *mut State;
-        let mut sum: u32 = 0;
         let mut i: std::ffi::c_uint = 0;
+        let mut count: u32 = 0;
+        let mut hiCnt: u32 = 0;
         let mut summFreq: u32 = (*(*p).min_context).union2.summ_freq as u32;
         if summFreq > (*p).range {
             summFreq = (*p).range;
         }
-        if (*s).symbol as std::ffi::c_int == symbol {
-            range_encoder_encode(p, 0 as std::ffi::c_int as u32, (*s).freq as u32, summFreq);
+        (*p).range = (*p).range / summFreq;
+        count = (*p).code / (*p).range;
+        hiCnt = count;
+        count = count.wrapping_sub((*s).freq as u32);
+        if (count as i32) < 0 as std::ffi::c_int {
+            let mut sym: u8 = 0;
+            range_decoder_decode(p, 0 as std::ffi::c_int as u32, (*s).freq as u32);
             while (*p).low ^ ((*p).low).wrapping_add((*p).range)
                 < (1 as std::ffi::c_int as u32) << 24 as std::ffi::c_int
                 || (*p).range < (1 as std::ffi::c_int as u32) << 15 as std::ffi::c_int && {
@@ -44,24 +52,30 @@ pub unsafe fn encode_symbol(p: *mut Ppmd8, symbol: std::ffi::c_int) {
                     1 as std::ffi::c_int != 0
                 }
             {
-                ((*(*p).stream.output).write).expect("non-null function pointer")(
-                    (*p).stream.output,
-                    ((*p).low >> 24 as std::ffi::c_int) as u8,
-                );
+                (*p).code = (*p).code << 8 as std::ffi::c_int
+                    | ((*(*p).stream.input).read).expect("non-null function pointer")(
+                        (*p).stream.input,
+                    ) as u32;
                 (*p).range <<= 8 as std::ffi::c_int;
                 (*p).low <<= 8 as std::ffi::c_int;
             }
             (*p).found_state = s;
+            sym = (*s).symbol;
             update1_0(p);
-            return;
+            return sym as std::ffi::c_int;
         }
         (*p).prev_success = 0 as std::ffi::c_int as std::ffi::c_uint;
-        sum = (*s).freq as u32;
         i = (*(*p).min_context).num_stats as std::ffi::c_uint;
         loop {
             s = s.offset(1);
-            if (*s).symbol as std::ffi::c_int == symbol {
-                range_encoder_encode(p, sum, (*s).freq as u32, summFreq);
+            count = count.wrapping_sub((*s).freq as u32);
+            if (count as i32) < 0 as std::ffi::c_int {
+                let mut sym_0: u8 = 0;
+                range_decoder_decode(
+                    p,
+                    hiCnt.wrapping_sub(count).wrapping_sub((*s).freq as u32),
+                    (*s).freq as u32,
+                );
                 while (*p).low ^ ((*p).low).wrapping_add((*p).range)
                     < (1 as std::ffi::c_int as u32) << 24 as std::ffi::c_int
                     || (*p).range < (1 as std::ffi::c_int as u32) << 15 as std::ffi::c_int && {
@@ -71,24 +85,28 @@ pub unsafe fn encode_symbol(p: *mut Ppmd8, symbol: std::ffi::c_int) {
                         1 as std::ffi::c_int != 0
                     }
                 {
-                    ((*(*p).stream.output).write).expect("non-null function pointer")(
-                        (*p).stream.output,
-                        ((*p).low >> 24 as std::ffi::c_int) as u8,
-                    );
+                    (*p).code = (*p).code << 8 as std::ffi::c_int
+                        | ((*(*p).stream.input).read).expect("non-null function pointer")(
+                            (*p).stream.input,
+                        ) as u32;
                     (*p).range <<= 8 as std::ffi::c_int;
                     (*p).low <<= 8 as std::ffi::c_int;
                 }
                 (*p).found_state = s;
+                sym_0 = (*s).symbol;
                 update1(p);
-                return;
+                return sym_0 as std::ffi::c_int;
             }
-            sum = sum.wrapping_add((*s).freq as u32);
             i = i.wrapping_sub(1);
             if !(i != 0) {
                 break;
             }
         }
-        range_encoder_encode(p, sum, summFreq.wrapping_sub(sum), summFreq);
+        if hiCnt >= summFreq {
+            return -(2 as std::ffi::c_int);
+        }
+        hiCnt = hiCnt.wrapping_sub(count);
+        range_decoder_decode(p, hiCnt, summFreq.wrapping_sub(hiCnt));
         let mut z: usize = 0;
         z = 0 as std::ffi::c_int as usize;
         while z < 256usize.wrapping_div(::core::mem::size_of::<usize>()) {
@@ -127,6 +145,7 @@ pub unsafe fn encode_symbol(p: *mut Ppmd8, symbol: std::ffi::c_int) {
             }
         }
     } else {
+        let s_0: *mut State = &mut (*(*p).min_context).union2 as *mut Union2 as *mut State;
         let prob: *mut u16 = &mut *(*((*p).bin_summ).as_mut_ptr().offset(
             *((*p).ns2index).as_mut_ptr().offset(
                 ((*(&mut (*(*p).min_context).union2 as *mut Union2 as *mut State)).freq as usize)
@@ -150,17 +169,17 @@ pub unsafe fn encode_symbol(p: *mut Ppmd8, symbol: std::ffi::c_int) {
                 .wrapping_add((*(*p).min_context).flags as std::ffi::c_int as std::ffi::c_uint)
                 as isize,
         ) as *mut u16;
-        let s_0: *mut State = &mut (*(*p).min_context).union2 as *mut Union2 as *mut State;
         let mut pr: u32 = *prob as u32;
-        let bound: u32 = ((*p).range >> 14 as std::ffi::c_int) * pr;
+        let size0: u32 = ((*p).range >> 14 as std::ffi::c_int) * pr;
         pr = pr.wrapping_sub(
             pr.wrapping_add(
                 ((1 as std::ffi::c_int) << 7 as std::ffi::c_int - 2 as std::ffi::c_int) as u32,
             ) >> 7 as std::ffi::c_int,
         );
-        if (*s_0).symbol as std::ffi::c_int == symbol {
+        if (*p).code < size0 {
+            let mut sym_1: u8 = 0;
             *prob = pr.wrapping_add(((1 as std::ffi::c_int) << 7 as std::ffi::c_int) as u32) as u16;
-            (*p).range = bound;
+            (*p).range = size0;
             while (*p).low ^ ((*p).low).wrapping_add((*p).range)
                 < (1 as std::ffi::c_int as u32) << 24 as std::ffi::c_int
                 || (*p).range < (1 as std::ffi::c_int as u32) << 15 as std::ffi::c_int && {
@@ -170,10 +189,10 @@ pub unsafe fn encode_symbol(p: *mut Ppmd8, symbol: std::ffi::c_int) {
                     1 as std::ffi::c_int != 0
                 }
             {
-                ((*(*p).stream.output).write).expect("non-null function pointer")(
-                    (*p).stream.output,
-                    ((*p).low >> 24 as std::ffi::c_int) as u8,
-                );
+                (*p).code = (*p).code << 8 as std::ffi::c_int
+                    | ((*(*p).stream.input).read).expect("non-null function pointer")(
+                        (*p).stream.input,
+                    ) as u32;
                 (*p).range <<= 8 as std::ffi::c_int;
                 (*p).low <<= 8 as std::ffi::c_int;
             }
@@ -182,6 +201,7 @@ pub unsafe fn encode_symbol(p: *mut Ppmd8, symbol: std::ffi::c_int) {
                 ((*s_0).successor_0 as u32 | ((*s_0).successor_1 as u32) << 16 as std::ffi::c_int)
                     as isize,
             ) as *mut std::ffi::c_void as *mut Context;
+            sym_1 = (*s_0).symbol;
             (*p).found_state = s_0;
             (*p).prev_success = 1 as std::ffi::c_int as std::ffi::c_uint;
             (*p).run_length += 1;
@@ -198,15 +218,16 @@ pub unsafe fn encode_symbol(p: *mut Ppmd8, symbol: std::ffi::c_int) {
             } else {
                 update_model(p);
             }
-            return;
+            return sym_1 as std::ffi::c_int;
         }
         *prob = pr as u16;
         (*p).init_esc = (*p).exp_escape[(pr >> 10 as std::ffi::c_int) as usize] as std::ffi::c_uint;
-        (*p).low = ((*p).low).wrapping_add(bound);
+        (*p).low = ((*p).low).wrapping_add(size0);
+        (*p).code = ((*p).code).wrapping_sub(size0);
         (*p).range = ((*p).range
             & !(((1 as std::ffi::c_int) << 7 as std::ffi::c_int + 7 as std::ffi::c_int) as u32)
                 .wrapping_sub(1 as std::ffi::c_int as u32))
-        .wrapping_sub(bound);
+        .wrapping_sub(size0);
         let mut z_0: usize = 0;
         z_0 = 0 as std::ffi::c_int as usize;
         while z_0 < 256usize.wrapping_div(::core::mem::size_of::<usize>()) {
@@ -228,17 +249,20 @@ pub unsafe fn encode_symbol(p: *mut Ppmd8, symbol: std::ffi::c_int) {
                 charMask[z_0.wrapping_add(6 as std::ffi::c_int as usize) as usize];
             z_0 = z_0.wrapping_add(8 as std::ffi::c_int as usize);
         }
-        *(charMask.as_mut_ptr() as *mut u8).offset((*s_0).symbol as isize) =
-            0 as std::ffi::c_int as u8;
+        *(charMask.as_mut_ptr() as *mut u8).offset(
+            (*(&mut (*(*p).min_context).union2 as *mut Union2 as *mut State)).symbol as isize,
+        ) = 0 as std::ffi::c_int as u8;
         (*p).prev_success = 0 as std::ffi::c_int as std::ffi::c_uint;
     }
     loop {
-        let mut see: *mut See = 0 as *mut See;
         let mut s_1: *mut State = 0 as *mut State;
-        let mut sum_0: u32 = 0;
-        let mut escFreq: u32 = 0;
+        let mut s2_0: *mut State = 0 as *mut State;
+        let mut freqSum: u32 = 0;
+        let mut count_0: u32 = 0;
+        let mut hiCnt_0: u32 = 0;
+        let mut freqSum2: u32 = 0;
+        let mut see: *mut See = 0 as *mut See;
         let mut mc: *mut Context = 0 as *mut Context;
-        let mut i_0: std::ffi::c_uint = 0;
         let mut numMasked: std::ffi::c_uint = 0;
         while (*p).low ^ ((*p).low).wrapping_add((*p).range)
             < (1 as std::ffi::c_int as u32) << 24 as std::ffi::c_int
@@ -249,10 +273,9 @@ pub unsafe fn encode_symbol(p: *mut Ppmd8, symbol: std::ffi::c_int) {
                 1 as std::ffi::c_int != 0
             }
         {
-            ((*(*p).stream.output).write).expect("non-null function pointer")(
-                (*p).stream.output,
-                ((*p).low >> 24 as std::ffi::c_int) as u8,
-            );
+            (*p).code = (*p).code << 8 as std::ffi::c_int
+                | ((*(*p).stream.input).read).expect("non-null function pointer")((*p).stream.input)
+                    as u32;
             (*p).range <<= 8 as std::ffi::c_int;
             (*p).low <<= 8 as std::ffi::c_int;
         }
@@ -262,124 +285,122 @@ pub unsafe fn encode_symbol(p: *mut Ppmd8, symbol: std::ffi::c_int) {
             (*p).order_fall = ((*p).order_fall).wrapping_add(1);
             (*p).order_fall;
             if (*mc).suffix == 0 {
-                return;
+                return -(1 as std::ffi::c_int);
             }
             mc = ((*p).base).offset((*mc).suffix as isize) as *mut std::ffi::c_void as *mut Context;
             if !((*mc).num_stats as std::ffi::c_uint == numMasked) {
                 break;
             }
         }
-        (*p).min_context = mc;
-        see = make_esc_freq(p, numMasked, &mut escFreq);
-        s_1 = ((*p).base).offset((*(*p).min_context).union4.stats as isize) as *mut std::ffi::c_void
-            as *mut State;
-        sum_0 = 0 as std::ffi::c_int as u32;
-        i_0 = ((*(*p).min_context).num_stats as std::ffi::c_uint)
+        s_1 =
+            ((*p).base).offset((*mc).union4.stats as isize) as *mut std::ffi::c_void as *mut State;
+        let mut num: std::ffi::c_uint = ((*mc).num_stats as std::ffi::c_uint)
             .wrapping_add(1 as std::ffi::c_int as std::ffi::c_uint);
+        let mut num2: std::ffi::c_uint = num.wrapping_div(2 as std::ffi::c_int as std::ffi::c_uint);
+        num &= 1 as std::ffi::c_int as std::ffi::c_uint;
+        hiCnt_0 = (*s_1).freq as u32
+            & *(charMask.as_mut_ptr() as *mut u8).offset((*s_1).symbol as isize) as u32
+            & (0 as std::ffi::c_int as u32).wrapping_sub(num);
+        s_1 = s_1.offset(num as isize);
+        (*p).min_context = mc;
         loop {
-            let cur: std::ffi::c_uint = (*s_1).symbol as std::ffi::c_uint;
-            if cur as std::ffi::c_int == symbol {
-                let low: u32 = sum_0;
-                let freq_0: u32 = (*s_1).freq as u32;
-                let mut num2: std::ffi::c_uint = 0;
-                if ((*see).shift as std::ffi::c_int) < 7 as std::ffi::c_int && {
-                    (*see).count = ((*see).count).wrapping_sub(1);
-                    (*see).count as std::ffi::c_int == 0 as std::ffi::c_int
-                } {
-                    (*see).summ = (((*see).summ as std::ffi::c_int) << 1 as std::ffi::c_int) as u16;
-                    let fresh0 = (*see).shift;
-                    (*see).shift = ((*see).shift).wrapping_add(1);
-                    (*see).count = ((3 as std::ffi::c_int) << fresh0 as std::ffi::c_int) as u8;
-                }
-                (*p).found_state = s_1;
-                sum_0 = sum_0.wrapping_add(escFreq);
-                num2 = i_0.wrapping_div(2 as std::ffi::c_int as std::ffi::c_uint);
-                i_0 &= 1 as std::ffi::c_int as std::ffi::c_uint;
-                sum_0 =
-                    sum_0.wrapping_add(freq_0 & (0 as std::ffi::c_int as u32).wrapping_sub(i_0));
-                if num2 != 0 as std::ffi::c_int as std::ffi::c_uint {
-                    s_1 = s_1.offset(i_0 as isize);
-                    loop {
-                        let sym0_0: std::ffi::c_uint =
-                            (*s_1.offset(0 as std::ffi::c_int as isize)).symbol as std::ffi::c_uint;
-                        let sym1_0: std::ffi::c_uint =
-                            (*s_1.offset(1 as std::ffi::c_int as isize)).symbol as std::ffi::c_uint;
-                        s_1 = s_1.offset(2 as std::ffi::c_int as isize);
-                        sum_0 = (sum_0 as std::ffi::c_uint).wrapping_add(
-                            (*s_1.offset(-(2 as std::ffi::c_int) as isize)).freq
-                                as std::ffi::c_uint
-                                & *(charMask.as_mut_ptr() as *mut u8).offset(sym0_0 as isize)
-                                    as std::ffi::c_uint,
-                        ) as u32 as u32;
-                        sum_0 = (sum_0 as std::ffi::c_uint).wrapping_add(
-                            (*s_1.offset(-(1 as std::ffi::c_int) as isize)).freq
-                                as std::ffi::c_uint
-                                & *(charMask.as_mut_ptr() as *mut u8).offset(sym1_0 as isize)
-                                    as std::ffi::c_uint,
-                        ) as u32 as u32;
-                        num2 = num2.wrapping_sub(1);
-                        if !(num2 != 0) {
-                            break;
-                        }
-                    }
-                }
-                if sum_0 > (*p).range {
-                    sum_0 = (*p).range;
-                }
-                range_encoder_encode(p, low, freq_0, sum_0);
-                while (*p).low ^ ((*p).low).wrapping_add((*p).range)
-                    < (1 as std::ffi::c_int as u32) << 24 as std::ffi::c_int
-                    || (*p).range < (1 as std::ffi::c_int as u32) << 15 as std::ffi::c_int && {
-                        (*p).range = (0 as std::ffi::c_int as u32).wrapping_sub((*p).low)
-                            & ((1 as std::ffi::c_int as u32) << 15 as std::ffi::c_int)
-                                .wrapping_sub(1 as std::ffi::c_int as u32);
-                        1 as std::ffi::c_int != 0
-                    }
-                {
-                    ((*(*p).stream.output).write).expect("non-null function pointer")(
-                        (*p).stream.output,
-                        ((*p).low >> 24 as std::ffi::c_int) as u8,
-                    );
-                    (*p).range <<= 8 as std::ffi::c_int;
-                    (*p).low <<= 8 as std::ffi::c_int;
-                }
-                update2(p);
-                return;
-            }
-            sum_0 = (sum_0 as std::ffi::c_uint).wrapping_add(
-                (*s_1).freq as std::ffi::c_uint
-                    & *(charMask.as_mut_ptr() as *mut u8).offset(cur as isize) as std::ffi::c_uint,
-            ) as u32 as u32;
-            s_1 = s_1.offset(1);
-            s_1;
-            i_0 = i_0.wrapping_sub(1);
-            if !(i_0 != 0) {
+            let sym0_0: std::ffi::c_uint =
+                (*s_1.offset(0 as std::ffi::c_int as isize)).symbol as std::ffi::c_uint;
+            let sym1_0: std::ffi::c_uint =
+                (*s_1.offset(1 as std::ffi::c_int as isize)).symbol as std::ffi::c_uint;
+            s_1 = s_1.offset(2 as std::ffi::c_int as isize);
+            hiCnt_0 = hiCnt_0.wrapping_add(
+                (*s_1.offset(-(2 as std::ffi::c_int) as isize)).freq as u32
+                    & *(charMask.as_mut_ptr() as *mut u8).offset(sym0_0 as isize) as u32,
+            );
+            hiCnt_0 = hiCnt_0.wrapping_add(
+                (*s_1.offset(-(1 as std::ffi::c_int) as isize)).freq as u32
+                    & *(charMask.as_mut_ptr() as *mut u8).offset(sym1_0 as isize) as u32,
+            );
+            num2 = num2.wrapping_sub(1);
+            if !(num2 != 0) {
                 break;
             }
         }
-        let mut total: u32 = sum_0.wrapping_add(escFreq);
-        (*see).summ = ((*see).summ as u32).wrapping_add(total) as u16;
-        if total > (*p).range {
-            total = (*p).range;
+        see = make_esc_freq(p, numMasked, &mut freqSum);
+        freqSum = freqSum.wrapping_add(hiCnt_0);
+        freqSum2 = freqSum;
+        if freqSum2 > (*p).range {
+            freqSum2 = (*p).range;
         }
-        range_encoder_encode(p, sum_0, total.wrapping_sub(sum_0), total);
-        let mut s2_0: *const State = ((*p).base).offset((*(*p).min_context).union4.stats as isize)
-            as *mut std::ffi::c_void as *mut State;
-        s_1 = s_1.offset(-1);
-        s_1;
-        *(charMask.as_mut_ptr() as *mut u8).offset((*s_1).symbol as isize) =
-            0 as std::ffi::c_int as u8;
+        (*p).range = (*p).range / freqSum2;
+        count_0 = (*p).code / (*p).range;
+        if count_0 < hiCnt_0 {
+            let mut sym_2: u8 = 0;
+            s_1 = ((*p).base).offset((*(*p).min_context).union4.stats as isize)
+                as *mut std::ffi::c_void as *mut State;
+            hiCnt_0 = count_0;
+            loop {
+                count_0 = count_0.wrapping_sub(
+                    (*s_1).freq as u32
+                        & *(charMask.as_mut_ptr() as *mut u8).offset((*s_1).symbol as isize) as u32,
+                );
+                s_1 = s_1.offset(1);
+                s_1;
+                if (count_0 as i32) < 0 as std::ffi::c_int {
+                    break;
+                }
+            }
+            s_1 = s_1.offset(-1);
+            s_1;
+            range_decoder_decode(
+                p,
+                hiCnt_0
+                    .wrapping_sub(count_0)
+                    .wrapping_sub((*s_1).freq as u32),
+                (*s_1).freq as u32,
+            );
+            while (*p).low ^ ((*p).low).wrapping_add((*p).range)
+                < (1 as std::ffi::c_int as u32) << 24 as std::ffi::c_int
+                || (*p).range < (1 as std::ffi::c_int as u32) << 15 as std::ffi::c_int && {
+                    (*p).range = (0 as std::ffi::c_int as u32).wrapping_sub((*p).low)
+                        & ((1 as std::ffi::c_int as u32) << 15 as std::ffi::c_int)
+                            .wrapping_sub(1 as std::ffi::c_int as u32);
+                    1 as std::ffi::c_int != 0
+                }
+            {
+                (*p).code = (*p).code << 8 as std::ffi::c_int
+                    | ((*(*p).stream.input).read).expect("non-null function pointer")(
+                        (*p).stream.input,
+                    ) as u32;
+                (*p).range <<= 8 as std::ffi::c_int;
+                (*p).low <<= 8 as std::ffi::c_int;
+            }
+            if ((*see).shift as std::ffi::c_int) < 7 as std::ffi::c_int && {
+                (*see).count = ((*see).count).wrapping_sub(1);
+                (*see).count as std::ffi::c_int == 0 as std::ffi::c_int
+            } {
+                (*see).summ = (((*see).summ as std::ffi::c_int) << 1 as std::ffi::c_int) as u16;
+                let fresh0 = (*see).shift;
+                (*see).shift = ((*see).shift).wrapping_add(1);
+                (*see).count = ((3 as std::ffi::c_int) << fresh0 as std::ffi::c_int) as u8;
+            }
+            (*p).found_state = s_1;
+            sym_2 = (*s_1).symbol;
+            update2(p);
+            return sym_2 as std::ffi::c_int;
+        }
+        if count_0 >= freqSum2 {
+            return -(2 as std::ffi::c_int);
+        }
+        range_decoder_decode(p, hiCnt_0, freqSum2.wrapping_sub(hiCnt_0));
+        (*see).summ = ((*see).summ as u32).wrapping_add(freqSum) as u16;
+        s_1 = ((*p).base).offset((*(*p).min_context).union4.stats as isize) as *mut std::ffi::c_void
+            as *mut State;
+        s2_0 = s_1
+            .offset((*(*p).min_context).num_stats as std::ffi::c_int as isize)
+            .offset(1 as std::ffi::c_int as isize);
         loop {
-            let sym0_1: std::ffi::c_uint =
-                (*s2_0.offset(0 as std::ffi::c_int as isize)).symbol as std::ffi::c_uint;
-            let sym1_1: std::ffi::c_uint =
-                (*s2_0.offset(1 as std::ffi::c_int as isize)).symbol as std::ffi::c_uint;
-            s2_0 = s2_0.offset(2 as std::ffi::c_int as isize);
-            *(charMask.as_mut_ptr() as *mut u8).offset(sym0_1 as isize) =
+            *(charMask.as_mut_ptr() as *mut u8).offset((*s_1).symbol as isize) =
                 0 as std::ffi::c_int as u8;
-            *(charMask.as_mut_ptr() as *mut u8).offset(sym1_1 as isize) =
-                0 as std::ffi::c_int as u8;
-            if !(s2_0 < s_1 as *const State) {
+            s_1 = s_1.offset(1);
+            s_1;
+            if !(s_1 != s2_0) {
                 break;
             }
         }
